Embed collection workflow in integrated tab and remove legacy entry files
